requirements_v2.txt
absl-py==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
dataframe-image==0.1.3
-e git+https://github.com/jinfagang/DCNv2_latest.git@fa9b2fd740ced2a22e0e7e913c3bf3934fd08098#egg=DCNv2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flatbuffers==1.12
fonttools==4.38.0
frozenlist==1.3.3
gast==0.3.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipython==8.6.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.1
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
keras==2.11.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.19.2
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.1
six==1.15.0
soupsieve==2.3.2.post1
stack-data==0.6.1
# tensorboard==2.11.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.4.0
# tensorflow-estimator==2.4.0
# termcolor==1.1.0
# tifffile==2022.10.10
# tinycss2==1.2.1
# torch==1.8.0+cu111
# torchaudio==0.8.0
# torchvision==0.9.0+cu111
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==3.7.4.3
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.12.1
XlsxWriter==3.0.3
yacs==0.1.8
yarl==1.8.1
zipp==3.10.0
